requirements.txt
selenium==4.1.3
Faker
requests
python-dotenv
pyperclip
openpyxl
oauth2client
google-api-python-client
google-auth-oauthlib
allure-behave==2.8.11
allure-python-commons==2.8.11
atomicwrites==1.3.0
attrs==19.3.0
behave==1.2.6
behave-generator==0.1.0
twirp==0.0.2
uvicorn==0.12.2
Appium-Python-Client==1.0.2
googletrans~=3.0.0
IMAPClient~=2.1.0